chore(coverage): Get the correct coverage numbers on all relevant files
Removes unnecessary __coverage__ plugin by using nyc\'s handy --all flag - this requires the\nbabel-register module to transpile things correctly for istanbul to instrument things. There also appears to be a bug with nyc checking the coverage stats in the package.json stanza, so passing the numbers directly on the command line for now.

The coverage has obviously dropped quite a lot, so a new threshold should be set and tests need to be written :)